Keep the build caches between builds
When rustflags change all build artifacts are invalidated, because they
are present in the build of all dependencies. This PR keeps the wasm
builds and the native builds separate allowing them to maintain their
caches. This may break some people's workflows if they're manually
building and running the integration tests.