Make the key version mandatory
We'll update key versions when we increase the security of our system,
or periodically lower the amount of assets accessible to former
validator cohorts. New key versions mean new keys.

Previously this field was optional, but if we leave it optional we have
to make a decision to either interpret None as 0, the least secure key
version, or continually update it to the latest one, changing peoples
keys from under them.

Since there isn't a good answer I feel it's better to make them
mandatory, so people at least have to think about them a bit.