Lowered the minimum gas limit and upped the depth
Observationally 30 transactions uses about 160 TG of Gas, so 250 should
be plenty. The protocol team says that this should come down soon.

The current gas limit makes it impossible for people to call the sign
method using another smart contract.

In the future we should abort when the remaining gas in the call starts
to get low to avoid the risk of an out of gas panic causing the system
to not clean up after itself.